rename ios project and update its dependencies
