+++ Examples/ios-math-solver/Sources/Assets.xcassets/AccentColor.colorset/Contents.json
