+++ Examples/ios-math-solver/Sources/Assets.xcassets/AppIcon.appiconset/Contents.json
