+++ Examples/ios-math-solver/Sources/Assets.xcassets/Contents.json
