+++ Examples/ios-math-solver/Sources/AuthenticationManager.swift
