+++ Examples/ios-math-solver/Sources/Content.swift
