+++ Examples/ios-math-solver/Sources/ContentView.swift
