+++ Examples/ios-math-solver/Sources/MathSolverApp.swift
