+++ Examples/ios-math-solver/Sources/Preview Content/Preview Assets.xcassets/Contents.json
