+++ Examples/ios-math-solver/Sources/math.entitlements
